Colorize compiler error/warning messages
If you've been mainly working with gcc during the past decade or so,
it has been outputting compiler warning and errors in colorized form,
if the output was a terminal.  Colorizing greatly helps to differentiate
errors from warnings.

Particularly since the TITAN compiler has (to my knowledge) no way to configure
which warnings to emit an which not, it's very difficult for the human
reader to pick out one 'error:' among potentially hundreds of 'warning:'
lines.  That's even more the case as the 'error' and 'warninig' is
printed after the variable-length location (file/line) and hence not
aligned in some column.

This patch adds output coloring to the TITAN compiler output, using the
same colors as gcc default colors.  This way, error messages are red,
warning messages are purple, and everything else is "default".

Signed-off-by: Harald Welte <laforge@gnumonks.org>
Change-Id: Ia5e42466ae3f964ce1bbaa159ae4f9f58259c616