Update GCC version check
At the beginning the GCC (and clang) used the versioning scheme:
<major>.<minor>.<patch>

The <patch> part is ignored by the version check as the patch releases are
compatible.

Both the GCC and the Clang dropped the minor releases at the release of GCC5 & Clang5
They garantee the API & ABI compatibility within a major version, so only the
major version should be taken account.

Change-Id: I73447b7f994b0b9eb5c536a49d4fe537bbd30969