Remove leading zero warning from str2float (bug 566028)
The standard explicitly allow the leading zero for str2float
Leading zero warning is not needed as it is allowed.

C.1.29  Character string to float
str2float(in charstring invalue) return float This function
converts a charstring comprising a number into a float value.
The format of the number in the charstring shall follow rules
in clause 6.1.0, items a) or b) with the following exceptions:

leading zeros are allowed;

EXAMPLE:

str2float("12345.6") // is the same as str2float("123.456E+02")

Change-Id: Ia3a81017b234dae493cd33ccc4ce3ea124727be2
Signed-off-by: Gabor Szalai <gabor.szalai@ericsson.com>