Reduce the float printout precision (bug 566118)
The 17 digit precision was too precise for the float.
The 17 significant digit ensured that the parsed back value is
the same as the original, but the currently used print out function
introduced rounding errors, which renders the last digit useless.

Further development and experiment is needed, because neither functionality
provided by the c++ library prints the float as expected by the TTCN standard.

Change-Id: I328e8ed423c7d96f2ee46e5453e090a1c3fda6a8
Signed-off-by: Gabor Szalai <gabor.szalai@ericsson.com>